Implement ObservablePromise
It's a crossover of a Promise and an Observable. We need it as an abstract concept to describe an async process that could be a promise or an observable. We need to return something synchronously, while we will only know if a worker function returns a promise or an observable once we receive its async message.